Adding stanza filters to our xmpp lib.
Two default filters were created. One to handle some events
like presence, messages and iqs received by the client. The other
was the id filter. This filter is heavly based on the libtc filter
implementation, based on GHook and GHookList glib data types.
Maybe in the future a simple linked list should be used.

A known problem here is that all hook lists are not being freed. It
will be done soon. And we should write a way to remove an added
filter. It is specialy important to the id filter, since we need to
add the filter, handle the message and then remove it from the
hook list.

In this commit some g_messages are replaced by fprintfs.