Adding an atom implementation.
Now libjarvis couns on an atom implementation. All xml stuff here
is handled with libiksemel because of it is easier and smaller
than things like libexpat or libxml2. There is no Feed implementation
yet and a serious bug is that JAtomEntry doesn't handle an ID
properly. This should be fixed soon.