First pieces of the python bindings for libjarvis.
We wrote a program to scan header files (it is not
so smart, but works with well defined standards in
header files) this program is divided in two parts,
the scanner and writer. The scanner can be used to
collect data for any header file that uses our
standards. The writer will be divided by language,
so by this way, we will be able to generate bindings
for any language.