Complete rewrite of binding system.
Now we have a dir called bindings/defs that holds a json generator. The
scanner module inside that dir generates a json file with collected info. Since
almost all common languages today has a json implementation, it will make
these informations available for everyone that wants to write a binding
for our library.

The building system was updated to look for python >= 2.5 and to
gen the json file and python binding modules. In the next few commits
we'll add the compilation support for the python binding.

A little regression was added here. Methods currently cannot be overriden,
but it is going to be fixed in the next commit.

A known bug serious in this implementation is about parameter modifiers,
when using optional parameter in some places, it inserts more than one "|"
to be parsed by PyArg_ParseTuple and it is breaking somethings.