Improvements in python binding writer.
Now it is possible to override a typedef and there is a return type to
the `void *' type. It is passed as a pure PyObject pointer if there is
something in the pointer, otherwise, None is returned.

This last change can help people make mistakes when getting objects
that are *not* python objects, so be carefull.