Improving our xmpp client and making it more robust.
Now, all fields in our JXmpp struct have getters and setters,
there is a way for get current available filters. More error handling was
added in j_xmpp_run, _j_xmpp_do_run and other places. The glib
dependency was purged, all tests for NULL stuff were purged, since
there is no other place in library have it. The gthread was replaced
by a pthread call.

Now our JXmpp struct is opaque and it was moved to the .c file.

In this commit I'm just adding -lpthread to the lib flags of the library, in the
future I'll test its presency in the configure.ac file.