Adding a little (and hammered) log data type.
This log lib has info, warn, debug, critical and error levels. A log instance
has a name and in the future it will be possible to register instances in
a register and find them globally.

This implementation already counts on a handler that is dispatched
right before a log message is being printed. It can stop the message
emission.