Fixing a little bug in optional parameter creation.
Python optional flag, received by `PyArg_ParseTuple' is not per
parameter. If a parameter is optional in a function, all of the next
params are going to be too. So, only one `|' (optional marker) should
be added.

This can bring some problems if this kind of standard is not
enforced. Maybe it will be needed to create a new macro to mark
nullable parameters. By this way we'll be able to make it different
from optional ones.