Adding overrides to the t_log_set_handler method.
Now, the python bindings user is able to to attach a callback to the
log handler. We needed to override the PyLogObject type definition
adding a callback and a parameter (python user_data) to it and
overriding t_log_set_handler properly to call a C callback that will
actually call the python callback.