Supporting "cyclic garbage collection" in some objects and other huge changes
The greatest change in this commit is the addition to the cyclic gc in types
that has python objects as attributes. To extend the attribute list the
override behaviour of a type was changed. Now, instead of the whole string,
the type generation waits for a list with the names of new attributes.
Currently we only support PyObject * attributes.

That was a huge change, but actually, was not the most important, since
currently I can't find any place that has these cyclic references happening,
but, that is the best way for handling these attributes references.

The most important change is about the override of filter methods, we have
a very (and unresolved) big deal here. The xmpp client class create and
expose a filter instance, when the filter instance is created, a C TXmpp
instance is passed as a user data parameter (void*). The problem in this
is that when calling the wrapper callback I didn't find a way to figure out
which type (if it is a PyObject* castable) to pass to the real python callback
function.

This commit holds a dirty test that creates a PyXmppObject from the 1st
parameter and a PyIksObject from the 3rd parameter. Obviously I'll try
to fix that thing in the next commits.