Testing for inconsistencies in python binding generator.
Two simple tests were added. First is when creating an object, the
generated constructor was retururning NULL values that was generating
a SystemError. Right now, I've assumed that a parameter was the cause
of the problem, but I know that it is not the only cause to C
constructors return NULL, this will be fixed someday...

The other fix was in the return type of each method that returns a
registered type. It, as the other problem, was not handling NULL
values.