Renaming type AtomAuthor to AtomPerson
In the Atom Format specification, there is a clear definition of
person and an entry author is just a person entity with the element
name set to `author'. So, I've changed the type name to fit this code
somewhere else that is waiting for a Person.

Now the to_string() and to_iks() methods are receiving an argument,
the element name. This is an example of the new API (in python):

  >>> from taningia import AtomPerson
  >>> ap = AtomPerson('Lincoln, 'lincoln@minaslivre.org')
  >>> ap.to_string('author')
  '<author><name>Lincoln</name><email>lincoln@minaslivre.org</email></author>'
  >>> ap.to_string('hacker')
  '<hacker><name>Lincoln</name><email>lincoln@minaslivre.org</email></hacker>'