Adding a basic error type to store error messages and codes.
This is not the best way for handling errors, but I really don't want to
have to pass an error instance in all functions that `raises' errors or
need to create a `context' with error handling.

In this way, an error object is created in each instance of the object that
actually uses it and a _get_error method should be provided to the
user of the object access the error instance.