Implementing the function to load an atom entry from a file.
Instead of making it a static method (or an alternative constructor) I
choosen to make it a simple method of an entry that can be called
after instantiating an empty entry.

The next step is use the TError object to handle errors in this library,
in generation, saving, parsing and loading steps.