Adding the `published' field in AtomEntry class.
It was a bit harder than the normal because `to_iks()' and `set_from_file()'
methods needed to be updated too. Other heavy change to make it work
better is the add of `time_t' automatic interpretation to our code gen.
When doing this, a little error was found when reading the storaged date
with `gmtime()'. Since we didn't read it with gmt info, we should not use it
to interpret the date object, so it was changed to `localtime()'.

Since published is a non required field in the Atom RFC, we needed
to add a new parameter modifier: '_nullable_' that handles fields required
but that can receive null values.