Purging node prefix building feature.
Since ejabberd 2.1 ships the flat node plugin by default (actually, we
were going to use it in 2.0 too) we'll not want to build the home based
tree structure as default anymore. But it is still possible to set a prefix
in the pubsub context to all nodes.

A little bug in concatenation of node_prefix and node_name was fixed
too.