[python-bindings] Separating libraries in different python modules.
This will give us a better python API and will improve code organization. This
commit introduces (atcually exposes) a circular dependency problem
between filter and xmpp python modules. Filter should not deppend on
xmpp, but this will not be fixed in this commit, so we got a little regression
when getting user data from a Filter object.

The TODO was file updated too.