Improving Iri API consistence by exposing Iri.set_from_string() method.
In some other places, like atom.{Entry,Feed} classes, we can see
`_set_from_*()' methods that are called *after* instantiating the class.
And in the Iri class, we were hiding the _set_from* method and calling
it from the constructor and returning null if something wrong happens.

Another advantage of doing this that it will be easier to improve our
error handling strategy. All objects that `raises' errors will have a
`get_error()' method and the user will be able to handle the error
and then free the instance.