Handling errors in atom and iri implementations with TError type.
Now, iri.Iri and atom.{Entry,Feed} functions that generates errors are
using the (new) TError interface. The error instance is stored as an
attribute of the class and is set when an error happens. To catch the
error, you can use the `_get_error()' method of the instance and then
access the name, code and message error attributes.

To use this interface is suggested to create an enum with all
available error types in the module and then use the t_error_set_full
method to fill all attributes at once.

src/xmpp.c and iri-example.c were updated to to use the new error
interface.