Heavy changes in the API and library standards.
Some of our basic standards of coding style, name conventions
indentation and stuff were changed in this commit. This was done
for so many reasons but the strongest was that we were typing too
much and names were a bit ugly =P.

To improve the API, functions like atom_feed_get_entries,
atom_entry_get_authors and others should return a list
instead of using out parameters. So, to fix this problem,
a basic implementation of a doubly linked list was added
to our code base. It is actually almost enougth to what
we want. Maybe after adding a sort algorithm to it we can
say that it is complete.

Another very good change in this commit is the by-function
documentation. All public functions are now documented.
These docstrings are special. We are providing some info
about how the function works, like if it raises some error
or the list items type. Our plan is to use these annotations
in docstrings to provide info to write bindings. This will
make it required to provide a docstring of a public function
to have it binded. But I think it is not a problem.

Another good stuff is that, since our docstrings are following
a basic standard, it will be easy to write a document extract
tool and generate a pretty API reference documentation.

The xmpp client class was heavely changed too. Its API was
simplified and now we have some regressions in it.

The main thing changed in the xmpp client was the way thet
we handle filters. Tests are being made to try iksfilter
structure and abandon our filter implementation that is
kinda buggy and incomplete.

Another important point that we will explore in the next
commits in xmpp client is to provide a way to notify
the user that something has happened. Like in the auth
process, we should inform to the API user when the
auth did not work and when it did work. We will make
some more tests about it. In this commit two hardcoted
hooks are being added.

PS.: This commit breaks all examples (only list example
works, because it was just implemented) and python bindings.
It will be fixed in this branche before merging with master.