Improving event dispatcher in xmpp client.
Now, instead of having only two hardcoded events, we have
a hash table storing it, so new events can be created if
it is needed.

The way of dispatching it was changed too. We have now a
delegate type that is received by the event_connect method.

The only problem in this implementation is inform to the
binding writer program which type of parameter is being
passed in each dispatched event. I think it should be
informed in an annotation in a docstring, so we will have
to extend the docstring parser (not commited yet) to
catch info about events.

Maybe we should generalize this event system and make
it a new separated module to allow user created
objects access and use it...