[python-bindings] Updating bindings to reflect API changes.
Since we have made huge changes in the library api and
coding standard, we needed another huge change in the
binding generation. Now aditional information of C
stuff can be provided from a docstring.  Currently, only
methods are using it but it can be extended to any
C type.

A very important improvment in this commit was the
fix automatic generation of functions that returns
lists. An annotation added to the method docstring
inform which data type is being stored in the list.

Note that methods with no docstrings will *NOT* be
reached by the scanner and thus not be exposed
in any binding.