[python-bindings] Fixes in documentation generation.
First of all, the docstring description field starts with
two \n chars but it can actually contain other \n\n. The
split that gets it was breaking it in more pieces if more
than one \n\n found.

The other improvement was simple but avoided some syntax
erros. Description was not escaping quotes.