Fixing a segfault in xmpp client.
A segfault was happening when creating a new client object passing
NULL to the `host' parameter. It was trying to get the server info
from the iksid object. But some commits ago I moved iksid
instantiation to the connect function... Since I saw it is wrong, I
moved it back to the constructor. To do it, I needed to create a new
field in the client, an iks stack to store the iksid. It is freed in
client destructor funtion.