Adding the Tag class to our iri module.
There are two important changes in this commit. The first and
obivous is that now our lib supports tag uri parsing and generation.
The second and very important thing is that the tag class
encapsulate an iri object inside it.

To do the second thing, we needed to expose both structs in the
.h file but it doesn't mean that anyone should access any field
directly.

This will be the standard for reusing code in the rest of the library.
More documentation tags were added here. Now the @type tag
can have the `class' value to describe a class definition. It should
be put near (preferable above) the struct of the type. These
changes were not implemented in the code scanner yet.