Removing glib dependency from taningia.
Now, we are free from the glib dependency. The last two pieces
are being removed in this commit. The first one, GHashTable from
the xmpp_client event machinery, that was replaced by a small
but nice implementation shamelesly stolen from the janson json
lib. And the g_time_val_to_iso8601, rewriten.