Fixed an invalid memory access, freeing data when removing a list item
Changing `ta_list_remove()' method to receive a new parameter, the
data_free_cb. It is a delegate (`ta_free_func_t') that, if given,
frees data field of a list item. A new header (mem.h) was added to
hold this delegate but I think more memory related things can fit
there.

The invalid memory access fixed was happening when removing the first
element of a list.