Accepting NULL host information when connecting
Until now, connection was being done by using `iks_connect_via()'
function, that receives a host to connect to. This way, we just don't
give the change to iksemel look for SRV records. Now, if the host is
NULL, we call `iks_connect_tcp()' instead.

The problem is that we are using libiksemel 1.2 and it has some
problems when connecting to localhost (with SRV records properly
configured). Trying iksemel 1.5 we get a worse problem, it connects
perfectly but fails in `iks_recv()'. So, it is not a taningia
problem. I'll investigate it later in iksemel or just file a bug
report.