Fixing a memory leak in `ta_list_insert()' and small memory errors
When inserting at the end of the list, append is called but the
already allocated new node was not being freed in insert. The
`ta_list_free()' was referencing a possibly already freed element.

The `list-example.c' file received some fixes too, like in the list
reverse example. Some comments were added too.