Improving configure.ac script.
There were some checks missing, found when running autoscan program.
Now it is possible to know if the system compiling our library has
some important things like `strndup()` func, a GNU extension.

The next step is to provide an alternative this and other extensions
when it is not found in the system.