Adding a new method to the xmpp client: `_send_and_filter()'
This method is very close to the `_send()' one. The important
difference is that it receives a callback that will be fired when the
server answers the sent message.

It is important to remember that this only works if the sent message
has an id attribute. Otherwise you'll get a segfault =/ because of a
`strdup()' against a NULL pointer.