Implementing refcount stuff in xmpp client.
Different from other already changed types, the xmpp client
did not publish its struct, since we do not want to expose
the internal hashtable implementation. The problem here is that
this `class` is not `inheritable` because of this detail.

I am not being careful here because our main goal is to implement
an easy to maintain refcount structure and we are achieving it.
In another branch, focused in an object system, this should be
fixed.