Heavy break in the pubsub API
This, in fact, is not true. We never had a pubsub API. The
first attemp was to create two objects, Pubsub and PubsubNode to
manage general pubsub stuff and pubsub node respectivly.

After some work in other test projects, like psbrowser, I realized
that it was ugly and not functional. This new attemp makes every
method becomes a function and gets rid of the two objects.