Some work on the ownership of known objects as parameters
Before merging the refcount branch into master, all objects were
allocated by the caller of a function of the API but, almost all
times, freed by the called function/object. This commit changes this
behaviour.

Now, when you pass a known object (ta_*_t, but ta_list_t) to a
function/method of the public API, the refcount of this object will be
incremented. IOW now the responsibility for creating *AND freeing* the
object passed to this API is on the caller.