Renaming srv_target to srv_resolver
We'll need both structutres, a resolver and a target. But at this time,
I've most worked in the resolver and, the target would be different from
what have now. The _resolver_query_domain() method will return a list
of targets. I'll start to write the target structure again when I start
to work in that method.