Refactoring the iksemel client hook
Instead of maintaining a huge method, the hook was split into smaller
and easier to maintain pieces.

Also, flags for sasl and tls were added to the client. The sasl one
is hardcoded to true and tls is hardcoded to false. This will be
improved soon.

Another thing is that now we are sending the presence, which means
the resource will be online as soon as the xmpp client connects