Compiling with -std=c99 instead of -ansi
This change actually caught a bug in the check_buf.c file. We were not
adding the test_buf_dealloc function to the test case.

Also, we got rid of the `foreign` param in the `AM_INIT_AUTOMAKE` call,
so we had to add the AUTHORS and the ChangeLog file and link the README
file to the README.md file.